+++ src/click/core.py
@@ -142,6 +142,9 @@ class ParameterSource(enum.Enum):
 
     .. versionchanged:: 8.0
         Use :class:`~enum.Enum` and drop the ``validate`` method.
+
+    .. versionchanged:: 8.0
+        Added the ``PROMPT`` value.
     """
 
     COMMANDLINE = enum.auto()
@@ -152,6 +155,8 @@ class ParameterSource(enum.Enum):
     """Used the default specified by the parameter."""
     DEFAULT_MAP = enum.auto()
     """Used a default provided by :attr:`Context.default_map`."""
+    PROMPT = enum.auto()
+    """Used a prompt to confirm a default or provide a value."""
 
 
 class Context:
@@ -277,10 +282,6 @@ class Context:
         #: Map of parameter names to their parsed values. Parameters
         #: with ``expose_value=False`` are not stored.
         self.params = {}
-        # This tracks the actual param objects that were parsed, even if
-        # they didn't expose a value. Used by completion system to know
-        # what parameters to exclude.
-        self._seen_params = set()
         #: the leftover arguments.
         self.args = []
         #: protected arguments.  These are arguments that are prepended
@@ -737,8 +738,12 @@ class Context:
 
         :param name: The name of the parameter.
         :rtype: ParameterSource
+
+        .. versionchanged:: 8.0
+            Returns ``None`` if the parameter was not provided from any
+            source.
         """
-        return self._parameter_source[name]
+        return self._parameter_source.get(name)
 
 
 class BaseCommand:
@@ -1283,7 +1288,11 @@ class Command(BaseCommand):
                 if (
                     not isinstance(param, Option)
                     or param.hidden
-                    or (not param.multiple and param in ctx._seen_params)
+                    or (
+                        not param.multiple
+                        and ctx.get_parameter_source(param.name)
+                        is ParameterSource.COMMANDLINE
+                    )
                 ):
                     continue
 
@@ -1779,6 +1788,15 @@ class Parameter:
         be removed in 8.1, until then it will be wrapped to match the
         new requirements.
 
+    .. versionchanged:: 8.0
+        For ``multiple=True, nargs>1``, the default must be a list of
+        tuples.
+
+    .. versionchanged:: 8.0
+        Setting a default is no longer required for ``nargs>1``, it will
+        default to ``None``. ``multiple=True`` or ``nargs=-1`` will
+        default to ``()``.
+
     .. versionchanged:: 7.1
         Empty environment variables are ignored rather than taking the
         empty string value. This makes it possible for scripts to clear
@@ -1926,16 +1944,20 @@ class Parameter:
             value = ctx.lookup_default(self.name)
             source = ParameterSource.DEFAULT_MAP
 
-        if value is not None:
-            ctx.set_parameter_source(self.name, source)
+        if value is None:
+            value = self.get_default(ctx)
+            source = ParameterSource.DEFAULT
 
-        return value
+        return value, source
 
     def type_cast_value(self, ctx, value):
         """Given a value this runs it properly through the type system.
         This automatically handles things like `nargs` and `multiple` as
         well as composite types.
         """
+        if value is None:
+            return () if self.multiple or self.nargs == -1 else None
+
         if self.type.is_composite:
             if self.nargs <= 1:
                 raise TypeError(
@@ -1943,14 +1965,17 @@ class Parameter:
                     f" been set to {self.nargs}. This is not supported;"
                     " nargs needs to be set to a fixed value > 1."
                 )
+
             if self.multiple:
-                return tuple(self.type(x or (), self, ctx) for x in value or ())
-            return self.type(value or (), self, ctx)
+                return tuple(self.type(x, self, ctx) for x in value)
+
+            return self.type(value, self, ctx)
 
         def _convert(value, level):
             if level == 0:
                 return self.type(value, self, ctx)
-            return tuple(_convert(x, level - 1) for x in value or ())
+
+            return tuple(_convert(x, level - 1) for x in value)
 
         return _convert(value, (self.nargs != 1) + bool(self.multiple))
 
@@ -1975,12 +2000,6 @@ class Parameter:
     def full_process_value(self, ctx, value):
         value = self.process_value(ctx, value)
 
-        if value is None and not ctx.resilient_parsing:
-            value = self.get_default(ctx)
-
-            if value is not None:
-                ctx.set_parameter_source(self.name, ParameterSource.DEFAULT)
-
         if self.required and self.value_is_missing(value):
             raise MissingParameter(ctx=ctx, param=self)
 
@@ -1988,8 +2007,12 @@ class Parameter:
         if (
             not ctx.resilient_parsing
             and self.nargs > 1
-            and self.nargs != len(value)
             and isinstance(value, (tuple, list))
+            and (
+                any(len(v) != self.nargs for v in value)
+                if self.multiple
+                else len(value) != self.nargs
+            )
         ):
             were = "was" if len(value) == 1 else "were"
             ctx.fail(
@@ -2002,44 +2025,45 @@ class Parameter:
     def resolve_envvar_value(self, ctx):
         if self.envvar is None:
             return
+
         if isinstance(self.envvar, (tuple, list)):
             for envvar in self.envvar:
                 rv = os.environ.get(envvar)
-                if rv is not None:
+
+                if rv:
                     return rv
         else:
             rv = os.environ.get(self.envvar)
 
-            if rv != "":
+            if rv:
                 return rv
 
     def value_from_envvar(self, ctx):
         rv = self.resolve_envvar_value(ctx)
+
         if rv is not None and self.nargs != 1:
             rv = self.type.split_envvar_value(rv)
+
         return rv
 
     def handle_parse_result(self, ctx, opts, args):
         with augment_usage_errors(ctx, param=self):
-            value = self.consume_value(ctx, opts)
+            value, source = self.consume_value(ctx, opts)
+            ctx.set_parameter_source(self.name, source)
+
             try:
                 value = self.full_process_value(ctx, value)
+
+                if self.callback is not None:
+                    value = self.callback(ctx, self, value)
             except Exception:
                 if not ctx.resilient_parsing:
                     raise
+
                 value = None
-            if self.callback is not None:
-                try:
-                    value = self.callback(ctx, self, value)
-                except Exception:
-                    if not ctx.resilient_parsing:
-                        raise
 
         if self.expose_value:
             ctx.params[self.name] = value
-
-        if value is not None:
-            ctx._seen_params.add(self)
 
         return value, args
 
@@ -2404,26 +2428,44 @@ class Option(Parameter):
 
         if rv is not None:
             return rv
+
         if self.allow_from_autoenv and ctx.auto_envvar_prefix is not None:
             envvar = f"{ctx.auto_envvar_prefix}_{self.name.upper()}"
-            return os.environ.get(envvar)
+            rv = os.environ.get(envvar)
+
+            if rv:
+                return rv
 
     def value_from_envvar(self, ctx):
         rv = self.resolve_envvar_value(ctx)
+
         if rv is None:
             return None
+
         value_depth = (self.nargs != 1) + bool(self.multiple)
+
         if value_depth > 0 and rv is not None:
             rv = self.type.split_envvar_value(rv)
+
             if self.multiple and self.nargs != 1:
                 rv = batch(rv, self.nargs)
+
         return rv
 
-    def full_process_value(self, ctx, value):
-        if value is None and self.prompt is not None and not ctx.resilient_parsing:
-            return self.prompt_for_value(ctx)
+    def consume_value(self, ctx, opts):
+        value, source = super().consume_value(ctx, opts)
 
-        return super().full_process_value(ctx, value)
+        # The value wasn't set, or used the param's default, prompt if
+        # prompting is enabled.
+        if (
+            source in {None, ParameterSource.DEFAULT}
+            and self.prompt is not None
+            and not ctx.resilient_parsing
+        ):
+            value = self.prompt_for_value(ctx)
+            source = ParameterSource.PROMPT
+
+        return value, source
 
 
 class Argument(Parameter):