+++ src/click/types.py
@@ -181,7 +181,7 @@ class StringParamType(ParamType):
                 else:
                     value = value.decode("utf-8", "replace")
             return value
-        return value
+        return str(value)
 
     def __repr__(self):
         return "STRING"
@@ -255,11 +255,9 @@ class Choice(ParamType):
         if normed_value in normed_choices:
             return normed_choices[normed_value]
 
-        self.fail(
-            f"invalid choice: {value}. (choose from {', '.join(self.choices)})",
-            param,
-            ctx,
-        )
+        one_of = "one of " if len(self.choices) > 1 else ""
+        choices_str = ", ".join(repr(c) for c in self.choices)
+        self.fail(f"{value!r} is not {one_of}{choices_str}.", param, ctx)
 
     def __repr__(self):
         return f"Choice({list(self.choices)})"
@@ -320,14 +318,19 @@ class DateTime(ParamType):
             return None
 
     def convert(self, value, param, ctx):
-        # Exact match
-        for format in self.formats:
-            dtime = self._try_to_convert_date(value, format)
-            if dtime:
-                return dtime
+        if isinstance(value, datetime):
+            return value
 
+        for format in self.formats:
+            converted = self._try_to_convert_date(value, format)
+
+            if converted is not None:
+                return converted
+
+        plural = "s" if len(self.formats) > 1 else ""
+        formats_str = ", ".join(repr(f) for f in self.formats)
         self.fail(
-            f"invalid datetime format: {value}. (choose from {', '.join(self.formats)})"
+            f"{value!r} does not match the format{plural} {formats_str}.", param, ctx
         )
 
     def __repr__(self):
@@ -341,7 +344,7 @@ class _NumberParamTypeBase(ParamType):
         try:
             return self._number_class(value)
         except ValueError:
-            self.fail(f"{value} is not a valid {self.name}", param, ctx)
+            self.fail(f"{value!r} is not a valid {self.name}.", param, ctx)
 
 
 class _NumberRangeBase(_NumberParamTypeBase):
@@ -495,7 +498,7 @@ class BoolParamType(ParamType):
     name = "boolean"
 
     def convert(self, value, param, ctx):
-        if isinstance(value, bool):
+        if value in {False, True}:
             return bool(value)
 
         norm = value.strip().lower()
@@ -506,7 +509,7 @@ class BoolParamType(ParamType):
         if norm in {"0", "false", "f", "no", "n", "off"}:
             return False
 
-        self.fail(f"{value!r} is not a valid boolean value.", param, ctx)
+        self.fail(f"{value!r} is not a valid boolean.", param, ctx)
 
     def __repr__(self):
         return "BOOL"
@@ -518,10 +521,15 @@ class UUIDParameterType(ParamType):
     def convert(self, value, param, ctx):
         import uuid
 
+        if isinstance(value, uuid.UUID):
+            return value
+
+        value = value.strip()
+
         try:
             return uuid.UUID(value)
         except ValueError:
-            self.fail(f"{value} is not a valid UUID value", param, ctx)
+            self.fail(f"{value!r} is not a valid UUID.", param, ctx)
 
     def __repr__(self):
         return "UUID"
@@ -610,11 +618,7 @@ class File(ParamType):
                     ctx.call_on_close(safecall(f.flush))
             return f
         except OSError as e:  # noqa: B014
-            self.fail(
-                f"Could not open file: {filename_to_ui(value)}: {get_strerror(e)}",
-                param,
-                ctx,
-            )
+            self.fail(f"{filename_to_ui(value)!r}: {get_strerror(e)}", param, ctx)
 
     def shell_complete(self, ctx, param, incomplete):
         """Return a special completion marker that tells the completion
@@ -818,37 +822,55 @@ class Tuple(CompositeParamType):
 
 
 def convert_type(ty, default=None):
-    """Converts a callable or python type into the most appropriate
-    param type.
+    """Find the most appropriate :class:`ParamType` for the given Python
+    type. If the type isn't provided, it can be inferred from a default
+    value.
     """
     guessed_type = False
+
     if ty is None and default is not None:
-        if isinstance(default, tuple):
-            ty = tuple(map(type, default))
+        if isinstance(default, (tuple, list)):
+            # If the default is empty, ty will remain None and will
+            # return STRING.
+            if default:
+                item = default[0]
+
+                # A tuple of tuples needs to detect the inner types.
+                # Can't call convert recursively because that would
+                # incorrectly unwind the tuple to a single type.
+                if isinstance(item, (tuple, list)):
+                    ty = tuple(map(type, item))
+                else:
+                    ty = type(item)
         else:
             ty = type(default)
+
         guessed_type = True
 
     if isinstance(ty, tuple):
         return Tuple(ty)
+
     if isinstance(ty, ParamType):
         return ty
+
     if ty is str or ty is None:
         return STRING
+
     if ty is int:
         return INT
-    # Booleans are only okay if not guessed.  This is done because for
-    # flags the default value is actually a bit of a lie in that it
-    # indicates which of the flags is the one we want.  See get_default()
-    # for more information.
-    if ty is bool and not guessed_type:
-        return BOOL
+
     if ty is float:
         return FLOAT
+
+    # Booleans are only okay if not guessed. For is_flag options with
+    # flag_value, default=True indicates which flag_value is the
+    # default.
+    if ty is bool and not guessed_type:
+        return BOOL
+
     if guessed_type:
         return STRING
 
-    # Catch a common mistake
     if __debug__:
         try:
             if issubclass(ty, ParamType):
@@ -856,7 +878,9 @@ def convert_type(ty, default=None):
                     f"Attempted to use an uninstantiated parameter type ({ty})."
                 )
         except TypeError:
+            # ty is an instance (correct), so issubclass fails.
             pass
+
     return FuncParamType(ty)